requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic[email]==2.5.3
python-multipart==0.0.6

# Existing dependencies
certifi==2024.8.30
charset-normalizer==3.3.2
decorator==4.4.2
filelock==3.16.1
fsspec==2024.9.0
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==2.1.5
more-itertools==10.5.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==2.0.2
openai-whisper==20231117
pillow==10.4.0
proglog==0.1.10
psutil==5.9.8
regex==2024.9.11
requests==2.32.3
sympy==1.13.2
tiktoken==0.7.0
torch==2.4.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
